refactor(output): extract query tests to sibling file
